Deleted repositories folder and moved logic to services
